Replace event loop with blocking code
